Revert "Add time datatype and input"
