+++ projectforge-rest/src/main/kotlin/org/projectforge/ui/UIDataType.kt
@@ -36,7 +36,6 @@ enum class UIDataType {
     PICTURE,
     STRING,
     TASK,
-    TIME,
     TIMESTAMP,
     TIMEZONE,
     USER
