+++ projectforge-rest/src/main/kotlin/org/projectforge/ui/UIDataTypeUtils.kt
@@ -28,7 +28,6 @@ import org.projectforge.business.task.TaskDO
 import org.projectforge.framework.persistence.user.entities.PFUserDO
 import java.math.BigDecimal
 import java.time.LocalDate
-import java.time.LocalTime
 import java.util.*
 
 class UIDataTypeUtils {
@@ -56,7 +55,6 @@ class UIDataTypeUtils {
                 TaskDO::class.java -> UIDataType.TASK
                 Locale::class.java -> UIDataType.LOCALE
                 TimeZone::class.java -> UIDataType.TIMEZONE
-                LocalTime::class.java -> UIDataType.TIME
                 else -> null
             }
         }
