+++ projectforge-webapp/src/components/base/dynamicLayout/components/input/DynamicInputResolver.jsx
@@ -5,7 +5,6 @@ import DynamicTaskSelect from '../select/task';
 import DynamicAutoCompletion from './DynamicAutoCompletion';
 import DynamicDateInput from './DynamicDateInput';
 import DynamicInput from './DynamicInput';
-import DynamicTimeInput from './DynamicTimeInput';
 import DynamicTimestampInput from './DynamicTimestampInput';
 
 // All types of 'INPUT' will be resolved here.
@@ -46,9 +45,6 @@ function DynamicInputResolver({ dataType, autoCompletionUrl, ...props }) {
             Tag = DynamicInput;
             additionalProps.type = 'password';
             break;
-        case 'TIME':
-            Tag = DynamicTimeInput;
-            break;
         default:
             return <span>{`${dataType} Input is not implemented.`}</span>;
     }
@@ -71,7 +67,6 @@ DynamicInputResolver.propTypes = {
         'DECIMAL',
         'NUMBER',
         'PASSWORD',
-        'TIME',
     ]).isRequired,
     autoCompletionUrl: PropTypes.string,
 };
